Python 3.6 Compatability
Python 3.6 requires dictionary.items() to be placed into a list in order to iterate over those items. There was also some non-obvious issue using a lambda with map instead of a 'simple' loop for highlighting the set of hexes using the Selection helper class. Finally, the sys.path.append("..") trick wasn't working on my machine. I made the import calls explicit with the assumption that the parent directory is named Hexy. Not ideal... but relative imports weren't working otherwise.