HexMap is no longer a subclass of dict
Moved the dictionary inside of the HexMap object and wrote a bunch of
helper methods to go with it. Those methods may not be all that useful,
but they will mimic the behavior of a dictionary. It can be iterated
upon, but does not support dict-specific operations such as `update` or
`pop`.